Correct headings on TOD params
The fields on this table were mislabled, but we confirmed that the
proper numbers are used by the model regardless of the field heading.
Cube lookups work by position and not by name which is very bad design,
but it means that this is does not affect model output. Resolve #1